Move GetImageForComponent into its own file
GetImageForComponent and the extension image logic will change.
Splitting this out will make future changes easier to track.
